Time: 153 ms (21.00%) | Memory: 55.5 MB (10.31%) - LeetSync

+++ 1753-path-with-minimum-effort/path-with-minimum-effort.cpp
@@ -39,7 +39,7 @@ class Solution {
 public:
     int minimumEffortPath(vector<vector<int>>& heights) {
         int low = 0;
-        int high = (int)1e6;
+        int high = 1e6;
     
         while (low < high) {
             int mid = (low + high) / 2;
